api/src/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==4.4.2
fastapi==0.114.1
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-cloud-videointelligence==2.13.5
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
moviepy==1.0.3
pillow==10.4.0
proglog==0.1.10
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
rsa==4.9
sniffio==1.3.1
starlette==0.38.5
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6